Atualização do exercício 096

+++ Ex096.py
@@ -8,6 +8,7 @@ def area(l, c):
     print(f'A área do retângulo é de {a}m².')
 
 
-largura = int(input('Digite a largura de um retângulo(m): '))
-comprimento = int(input('Digite o comprimento do retângulo(m): '))
+#Programa principal
+largura = float(input('Digite a largura de um retângulo(m): '))
+comprimento = float(input('Digite o comprimento do retângulo(m): '))
 area(largura, comprimento)
